refactor: verifier -> verifier_utils
