+++ packages/rust-verifier/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod verifier;
+pub mod verifier_utils;
