+++ packages/rust-verifier/src/verifier_utils.rs
